requirements.txt
obsws-python==1.7.1
paho-mqtt==2.1.0
python-dotenv==1.1.0
websocket-client==1.8.0
